FIX action objects.
This uses the new data entities.  The problem was the entities
  had hidden meanings from the order of the table keys.
The new formats uses hashes for table keys, so that they can be worked
  on directly without mental acrobatics.
Also, added an adjacency list to the graph, removed the variable list
  from the factor, and generally made this thing work.
Added some support tools, but I do have a brittle factory: a_to_d.csv.
There are a lot of calculations that are based on these support files
  not changing.  Unfortunately, either I don't have the stamina to
  build more-forgiving tests, or the nature of this kind of operation
  is just context specific.  What I'm confident of is that the
  observations are being translated into tables, factors and
  graphs in a reasonable way.